requirement.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==2.1.4
django-appconf==1.0.2
django-filter==2.1.0
django-imagekit==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djongo==1.2.31
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
openapi-codec==1.3.2
pilkit==2.0
Pillow==5.3.0
Pygments==2.3.1
pymongo==3.7.2
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1